requirements.txt
pip==20.1.1
setuptools==47.1.0
tensorflow==2.4.1
matplotlib==3.3.4
SSIM-PIL==1.0.12
scipy==1.6.1
